+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 3.2.5
+* `NEW` provide config docs in `LUA_LANGUAGE_SERVER/doc/`
 * `FIX` [#1149](https://github.com/sumneko/lua-language-server/issues/1149)
 * `FIX` [#1192](https://github.com/sumneko/lua-language-server/issues/1192)