+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 3.2.5
+`2022-6-9`
 * `NEW` provide config docs in `LUA_LANGUAGE_SERVER/doc/`
 * `FIX` [#1148](https://github.com/sumneko/lua-language-server/issues/1148)
 * `FIX` [#1149](https://github.com/sumneko/lua-language-server/issues/1149)